fix: refresh system variable (VF-000) (#283)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements VF-000**

### Brief description. What is this change?
Nico informed me that the `input.requestEnvelope.context.System` can actually change with critical information through the runtime of a skill. 
Right now we only initialize it at the very beginning of the session (LaunchRequest) and never change it after. With this, we will bump it every single interaction/turn.