fix: remove integrations service (PL-901) (#339)
This fully removes the integrations service. The standard behavior is for a unknown node to get captured by the `NextHandler`:
https://github.com/voiceflow/general-runtime/blob/4e455014623d48d788339953af62c4de2260c367/runtime/lib/Handlers/next.ts#L9-L16

However, this will cause breakage of the agent, as the integration node only has `node.success_id` or `node.fail_id`.

Instead there is now just a stub handler that routes it to the fail port, or whatever port.

To unblock infra, we have to remove `INTEGRATIONS_HANDLER_ENDPOINT` as a required env var.

Co-authored-by: Tyler Han <tylerhan97@gmail.com>